T6

+++ database/migrations/2022_09_01_062700_update_goods_type_id_to_requests_table.php
@@ -13,16 +13,24 @@ class UpdateGoodsTypeIdToRequestsTable extends Migration
      */
     public function up()
     {
-        Schema::table('requests', function (Blueprint $table) {
+        
+        if(Schema::hasTable('requests')){
+            if(!Schema::hasColumn('requests','goods_type_id')){
+                Schema::table('requests', function (Blueprint $table) {
+                $table->unsignedInteger('goods_type_id')->after('user_id')->nullable();
+                $table->foreign('goods_type_id')
+                ->references('id')
+                ->on('goods_types')
+                ->onDelete('cascade');
+                });
+            }
 
-            $table->unsignedInteger('goods_type_id')->after('user_id')->nullable();
-            $table->string('goods_type_quantity')->after('goods_type_id')->nullable();
-
-            $table->foreign('goods_type_id')
-                    ->references('id')
-                    ->on('goods_types')
-                    ->onDelete('cascade');
-                 });
+            if(!Schema::hasColumn('requests','goods_type_quantity')){
+                Schema::table('requests', function (Blueprint $table) {
+                  $table->string('goods_type_quantity')->after('goods_type_id')->nullable();
+                });
+            }
+        }
     }
 
     /**
